* src/gtk/gtk_widget.h (NUM_ENTRY_WIDTH): change width of entries.
* src/gtk/gtk_widget.c
(create_direction_entry): change width of the entry.
(show_color_sel, color_button_key_event): not used if the version of GTK+ is greater or equal to 3.4.

* src/gtk/ox11menu.c
(get_palette, set_palette): not used if the version of GTK+ is greater or equal to 3.4.

* src/gtk/x11dialg.c
(_set_color, _putobj_color): use "GtkColorChooser" if the version of GTK+ is greater or equal to 3.4.

* src/gtk/x11axis.c
(position_tab_create): change width of the "Adjust:" entry.

* src/gtk/x11opt.c
(ViewerDialogSetupItem, ViewerDialogClose): use "GtkColorChooser" if the version of GTK+ is greater or equal to 3.4.

* src/gtk/x11view.c
(PasteObjectsFromClipboard): use "gdk_window_get_device_position()" instead of "gtk_widget_get_pointer()" if the version of GTK+ is greater or equal to 3.4.
(ViewerWinSetup): add "GDK_SCROLL_MASK" and "GDK_SMOOTH_SCROLL_MASK".
(ViewerEvScroll): add case for "GDK_SCROLL_SMOOTH".



git-svn-id: https://svn.code.sf.net/p/ngraph-gtk/code/trunk@469 a1697baf-5ae0-492e-bcf4-638f3baa2ba1