* src/plugins.c (plugin_shell_exec): check security. (ngraph_plugin_shell_err_puts, ngraph_plugin_shell_puts, ngraph_plugin_shell_get_inst): added.
* src/otext.c
(test): type of argument of the field "save_config" is "void".

* src/osystem.c
(nsystem): type of argument of the field "temp_file" is "void".

* src/ostring.c
(ostring): type of argument of some fields are "void".

* src/osarray.c
(osarray): type of argument of some fields are "void".

* src/oregexp.c
(oregexp): type of argument of some fields are "void".

* src/oprm.c
(prm): type of argument of the field "load" is "void".

* src/omerge.c
(merge): type of argument of some fields are "void".

* src/omath.c
(math): type of argument of the field "clear" is "void".

* src/oiarray.c
(oiarray): type of argument of some fields are "void".

* src/ofile.c
(file2d): type of argument of some fields are "void".

* src/odarray.c
(odarray): type of argument of some fields are "void".

* src/oaxis.c
(draw_numbering, numbering): type of the last parameter is "struct narray *".
(draw_numbering): use "g_strdup_printf()" when the parameter "array" is not NULL.
(axis_get_numbering): type of field is "NSAFUNC".
(axis): type of the argument of fields "scale_push", "scale_pop", "clear", "adjust", "tight", "group_manager" and "get_numbering" are "void".

* src/oagrid.c
(agrid): type of "tight" field is "void (void)".



git-svn-id: https://svn.code.sf.net/p/ngraph-gtk/code/trunk@671 a1697baf-5ae0-492e-bcf4-638f3baa2ba1